테스트 코드 모듈화 #6
