+++ lab/src/utils/configs.ts
@@ -1,0 +1,2 @@
+export const USER_AGENT_CHROME =
+    'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/95.0.4638.69 Safari/537.36';
