--- lab/src/utils/makeObject.ts
@@ -1,7 +1,0 @@
-export function makeObject(name: string, value: number) {
-    return { name, value };
-}
-
-export function testMakeObject() {
-    console.log(makeObject('Sannim', 28));
-}
